refactor: 장소가 없는 경우 예외를 던지게 수정
