+++ backend/src/main/java/feedzupzup/backend/feedback/domain/FeedBackRepository.java
@@ -6,15 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 public interface FeedBackRepository extends JpaRepository<Feedback, Long> {
-
-    @Query("""
-            SELECT f
-            FROM Feedback f
-            WHERE (:cursorId IS NULL OR f.id < :cursorId)
-            ORDER BY f.id DESC
-            """)
-    List<Feedback> findPageByCursorIdOrderByDesc(final Long cursorId, final Pageable pageable);
-
     @Query("""
             SELECT f
             FROM Feedback f
