+++ backend/src/main/java/feedzupzup/backend/place/domain/Place.java
@@ -1,13 +1,15 @@
-package feedzupzup.backend.place;
+package feedzupzup.backend.place.domain;
 
 import feedzupzup.backend.global.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Place extends BaseTimeEntity {
