+++ backend/src/main/java/feedzupzup/backend/place/domain/PlaceRepository.java
@@ -1,0 +1,7 @@
+package feedzupzup.backend.place.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlaceRepository extends JpaRepository<Place, Long> {
+
+}
